requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.32
psycopg[binary]>=3.2.1
chromadb>=0.5.5
pandas>=2.2.0
pyyaml>=6.0.1
requests>=2.31.0
